requirements.txt
androguard==3.3.5
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
beautifulsoup4==4.8.1
boto3==1.10.2
botocore==1.13.2
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.6.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
docutils==0.14
fake-useragent==0.1.11
future==0.18.1
gpapi==0.4.3
GPlayCli==3.26
idna==2.8
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.15.1
jmespath==0.9.4
jupyter-client==5.3.4
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.4.1
matplotlib==3.1.1
mccabe==0.6.1
networkx==2.4
numpy==1.17.3
pandas==0.25.2
parse==1.12.1
parso==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pyaxmlparser==0.3.21
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
pyee==6.0.0
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.1
python-dateutil==2.8.0
pytoml==0.1.20
pytz==2019.3
pyzmq==18.1.0
requests==2.22.0
requests-html==0.10.0
s3transfer==0.2.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.10
toml==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.3.5
urllib3==1.24.3
w3lib==1.21.0
wcwidth==0.1.7
websockets==8.0.2
wrapt==1.11.1
